wip エラーなく動くように修正
