+++ frontend/package.json
@@ -27,7 +27,8 @@
         "prettier": "^2.8.1",
         "react": "18.2.0",
         "react-dom": "18.2.0",
-        "typescript": "4.9.4"
+        "typescript": "4.9.4",
+        "material-symbols": "^0.4.1"
     },
     "devDependencies": {
         "autoprefixer": "^10.4.13",
