+++ frontend/src/components/uiGroup/RawpConcept.tsx
@@ -1,4 +1,4 @@
-import DescribeH2 from '@/components/uiParts/DescribeH2';
+import DescribeH2 from '@/components/uiParts/Heading/DescribeH2';
 import type { NextPage } from 'next';
 const RawpConcept: NextPage = () => {
     return (
