+++ frontend/src/pages/description.tsx
@@ -1,0 +1,12 @@
+import Layout from '@/layouts/Layout';
+import type { NextPage } from 'next';
+const DescriptionPage: NextPage = () => {
+    return (
+        <Layout title="アルゴリズムの説明">
+            
+        </Layout>
+    );
+};
+
+export default DescriptionPage;
+
