+++ frontend/src/styles/globals.css
@@ -1,3 +1,64 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+:root{
+    --heder-height: 100px;
+    --footer-height: 120px;
+    --main-padding-top: 20px;
+    --main-max-width: 1300px;
+    --font-kiwi-maru: 'Kiwi Maru';
+}
+@media screen and (max-width: 640px) {
+    :root {
+        --heder-height: 60px;
+        --footer-height: 120px;
+        --main-padding-top: 20px;
+        --main-max-width: 1300px;
+    }
+}
+
+body {
+    scroll-behavior: smooth;
+}
+header {
+    height: var(--heder-height);
+    backdrop-filter: blur(4px);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+footer {
+    min-height: var(--footer-height);
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    box-shadow: inset 0px 4px 4px rgba(0, 0, 0, 0.25);
+}
+main {
+    padding: 1rem;
+    max-width: var(--main-max-width);
+    margin: 0 auto;
+    min-height: calc(100vh - var(--heder-height) - var(--footer-height));
+    padding-top: var(--main-padding-top);
+}
+
+/* マテリアルアイコン関連のCSS */
+.material-symbols-outlined.md-18 {
+    font-size: 18px;
+}
+.material-symbols-outlined.md-24 {
+    font-size: 24px;
+}
+.material-symbols-outlined.md-36 {
+    font-size: 36px;
+}
+.material-symbols-outlined.md-48 {
+    font-size: 48px;
+}
+.material-symbols-outlined.md-100 {
+    font-size: 100px;
+}
+.material-symbols-outlined.font-weight-100 {
+    font-weight: 100;
+}
